refactor(tests): move assert_eq to utils;
